requirements/base.txt
Django==5.1
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2==2.9.9
django-phonenumber-field==8.0.0
